Refactored Products and Product models to remove SQL into data layer.
