+++ refactor-me/Database/IDataLayer.cs
@@ -1,4 +1,5 @@
 ﻿using refactor_me.Models;
+using System;
 
 namespace refactor_me.Database
 {
@@ -6,5 +7,9 @@ namespace refactor_me.Database
     {
         Products LoadProducts(string where);
         Products LoadProductsByName(string name);
+
+        Product LoadProductById(string id);
+        void SaveProduct(Product product);
+        void DeleteProduct(Guid id);
     }
 }
