+++ refactor-me/Models/Product.cs
@@ -1,0 +1,36 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace refactor_me.Models
+{
+    public class Product
+    {
+        public Guid Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Description { get; set; }
+
+        public decimal Price { get; set; }
+
+        public decimal DeliveryPrice { get; set; }
+
+        [JsonIgnore]
+        public bool IsNew { get; }
+
+        public Product()
+        {
+            Id = Guid.NewGuid();
+            IsNew = true;
+        }
+
+        public Product(Guid id)
+        {
+            Id = id;
+            IsNew = false;
+        }
+    }
+}
