+++ refactor-me/Models/Products.cs
@@ -1,174 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Linq;
 using Newtonsoft.Json;
-using refactor_me.Database;
+using System.Data.SqlClient;
 
 namespace refactor_me.Models
 {
     public class Products
     {
         public List<Product> Items { get; private set; }
-
-        public Products()
-        {
-
-        }
-
+        
         public Products(IEnumerable<Product> items)
         {
             Items = items.ToList();
-        }
-    }
-
-    public class Product
-    {
-        public Guid Id { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        public decimal Price { get; set; }
-
-        public decimal DeliveryPrice { get; set; }
-        
-        [JsonIgnore]
-        public bool IsNew { get; }
-
-        public Product()
-        {
-            Id = Guid.NewGuid();
-            IsNew = true;
-        }
-
-        public Product(Guid id)
-        {
-            IsNew = true;
-            var conn = Helpers.NewConnection();
-            var cmd = new SqlCommand($"select * from product where id = '{id}'", conn);
-            conn.Open();
-
-            var rdr = cmd.ExecuteReader();
-            if (!rdr.Read())
-                return;
-
-            IsNew = false;
-            Id = Guid.Parse(rdr["Id"].ToString());
-            Name = rdr["Name"].ToString();
-            Description = (DBNull.Value == rdr["Description"]) ? null : rdr["Description"].ToString();
-            Price = decimal.Parse(rdr["Price"].ToString());
-            DeliveryPrice = decimal.Parse(rdr["DeliveryPrice"].ToString());
-        }
-
-        public void Save()
-        {
-            var conn = Helpers.NewConnection();
-            var cmd = IsNew ? 
-                new SqlCommand($"insert into product (id, name, description, price, deliveryprice) values ('{Id}', '{Name}', '{Description}', {Price}, {DeliveryPrice})", conn) : 
-                new SqlCommand($"update product set name = '{Name}', description = '{Description}', price = {Price}, deliveryprice = {DeliveryPrice} where id = '{Id}'", conn);
-
-            conn.Open();
-            cmd.ExecuteNonQuery();
-        }
-
-        public void Delete()
-        {
-            foreach (var option in new ProductOptions(Id).Items)
-                option.Delete();
-
-            var conn = Helpers.NewConnection();
-            conn.Open();
-            var cmd = new SqlCommand($"delete from product where id = '{Id}'", conn);
-            cmd.ExecuteNonQuery();
-        }
-    }
-
-    public class ProductOptions
-    {
-        public List<ProductOption> Items { get; private set; }
-
-        public ProductOptions()
-        {
-            LoadProductOptions(null);
-        }
-
-        public ProductOptions(Guid productId)
-        {
-            LoadProductOptions($"where productid = '{productId}'");
-        }
-
-        private void LoadProductOptions(string where)
-        {
-            Items = new List<ProductOption>();
-            var conn = Helpers.NewConnection();
-            var cmd = new SqlCommand($"select id from productoption {where}", conn);
-            conn.Open();
-
-            var rdr = cmd.ExecuteReader();
-            while (rdr.Read())
-            {
-                var id = Guid.Parse(rdr["id"].ToString());
-                Items.Add(new ProductOption(id));
-            }
-        }
-    }
-
-    public class ProductOption
-    {
-        public Guid Id { get; set; }
-
-        public Guid ProductId { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        [JsonIgnore]
-        public bool IsNew { get; }
-
-        public ProductOption()
-        {
-            Id = Guid.NewGuid();
-            IsNew = true;
-        }
-
-        public ProductOption(Guid id)
-        {
-            IsNew = true;
-            var conn = Helpers.NewConnection();
-            var cmd = new SqlCommand($"select * from productoption where id = '{id}'", conn);
-            conn.Open();
-
-            var rdr = cmd.ExecuteReader();
-            if (!rdr.Read())
-                return;
-
-            IsNew = false;
-            Id = Guid.Parse(rdr["Id"].ToString());
-            ProductId = Guid.Parse(rdr["ProductId"].ToString());
-            Name = rdr["Name"].ToString();
-            Description = (DBNull.Value == rdr["Description"]) ? null : rdr["Description"].ToString();
-        }
-
-        public void Save()
-        {
-            var conn = Helpers.NewConnection();
-            var cmd = IsNew ?
-                new SqlCommand($"insert into productoption (id, productid, name, description) values ('{Id}', '{ProductId}', '{Name}', '{Description}')", conn) :
-                new SqlCommand($"update productoption set name = '{Name}', description = '{Description}' where id = '{Id}'", conn);
-
-            conn.Open();
-            cmd.ExecuteNonQuery();
-        }
-
-        public void Delete()
-        {
-            var conn = Helpers.NewConnection();
-            conn.Open();
-            var cmd = new SqlCommand($"delete from productoption where id = '{Id}'", conn);
-            cmd.ExecuteReader();
         }
     }
 }
